Separate declaration and definition of consensus initialization
